#2650 Put classes in separate files; move client-side dp options to the right location
